Move Nanoc::Int::ActionProvider
